+++ src/libpython.cpp
@@ -431,9 +431,9 @@ bool import_numpy_api(bool python3, std::string* pError) {
 
   // check C API version
   // we aim to compile a single binary compatible with both numpy 2.x and 1.x
-  auto runtime_npy_version = PyArray_GetNDArrayCVersion();
-  if (NPY_VERSION_2 != runtime_npy_version &&
-      NPY_VERSION_1 != runtime_npy_version) {
+  PyArray_RUNTIME_VERSION = PyArray_GetNDArrayCVersion();
+  if (NPY_VERSION_2 != PyArray_RUNTIME_VERSION &&
+      NPY_VERSION_1 != PyArray_RUNTIME_VERSION) {
     std::ostringstream ostr;
     ostr << "incompatible NumPy binary version " << (int) PyArray_GetNDArrayCVersion() << " "
     "(expecting version " << (int) NPY_VERSION_2 << " or " << (int) NPY_VERSION_1 << ")";